Update move shareusertestdata
