+++ .eslintrc.json
@@ -8,7 +8,8 @@
     "plugin:react/recommended",
     "plugin:react-hooks/recommended",
     "plugin:@typescript-eslint/recommended",
-    "plugin:prettier/recommended"
+    "plugin:prettier/recommended",
+    "next"
   ],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
@@ -34,5 +35,10 @@
     "no-unused-vars": "off",
     "@typescript-eslint/no-unused-vars": "error",
     "arrow-body-style": ["error", "as-needed"]
+  },
+  "settings": {
+    "react": {
+      "version": "detect"
+    }
   }
 }
